+++ src/main/kotlin/com/dh/baro/core/dlt/FailedMessageRepository.kt
@@ -1,4 +1,4 @@
-package com.dh.baro.core.kafka
+package com.dh.baro.core.dlt
 
 import org.springframework.data.jpa.repository.JpaRepository
